Create new class for mock controller, amplifier, carrier, launcher, write new test for booster and launcher
